SnowyDune 13/10

+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/entity/CarRental.java
@@ -17,7 +17,7 @@ public class CarRental {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idCarRental;
+	private Long idCarRental;
 	
 	private String name;
