+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/entity/Forfait.java
@@ -21,7 +21,7 @@ public class Forfait {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idForfait;
+	private Long idForfait;
 	
 	private int numberForfait;
