+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/entity/Hotel.java
@@ -17,7 +17,7 @@ public class Hotel {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idHotel;
+	private Long idHotel;
 	
 	private Double priceDay;
