+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/entity/PaymentHistory.java
@@ -21,7 +21,7 @@ public class PaymentHistory {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idPayment;
+	private Long idPayment;
 	
 	private Double paymentAmount;
