+++ src/BackEnd/SnowyDune/src/main/java/org/iesalixar/poriad/entity/Role.java
@@ -17,7 +17,7 @@ public class Role {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int idRole;
+	private Long idRole;
 	
 	private String name;
